refactor: address review feedback
- drop redundant ids/titles in filters.json and settings.json where
  identical to options/values
- use unwrap_or_default() instead of unwrap_or(ContentRating::Unknown)
- move encode_value, encode_list, build_listing_url, and push_scroller
  from lib.rs to helpers.rs to keep lib.rs focused on the Source trait
  implementation
