+++ sources/en.ranobes/res/filters.json
@@ -53,8 +53,7 @@
 		"id": "languages",
 		"title": "Original Language",
 		"canExclude": true,
-		"options": ["Chinese", "Korean", "English", "Japanese"],
-		"ids": ["Chinese", "Korean", "English", "Japanese"]
+		"options": ["Chinese", "Korean", "English", "Japanese"]
 	},
 	{
 		"type": "multi-select",
@@ -63,43 +62,6 @@
 		"isGenre": true,
 		"canExclude": true,
 		"options": [
-			"Action",
-			"Adult",
-			"Adventure",
-			"Comedy",
-			"Drama",
-			"Ecchi",
-			"Fantasy",
-			"Game",
-			"Gender Bender",
-			"Harem",
-			"Historical",
-			"Horror",
-			"Josei",
-			"Martial Arts",
-			"Mature",
-			"Mecha",
-			"Mystery",
-			"Psychological",
-			"Romance",
-			"School Life",
-			"Sci-fi",
-			"Seinen",
-			"Shoujo",
-			"Shounen",
-			"Shounen Ai",
-			"Slice of Life",
-			"Smut",
-			"Sports",
-			"Supernatural",
-			"Tragedy",
-			"Wuxia",
-			"Xianxia",
-			"Xuanhuan",
-			"Yaoi",
-			"Yuri"
-		],
-		"ids": [
 			"Action",
 			"Adult",
 			"Adventure",
