+++ sources/en.ranobes/res/settings.json
@@ -7,43 +7,6 @@
 				"type": "multi-select",
 				"key": "hiddenGenres",
 				"title": "Hidden Genres",
-				"titles": [
-					"Action",
-					"Adult",
-					"Adventure",
-					"Comedy",
-					"Drama",
-					"Ecchi",
-					"Fantasy",
-					"Game",
-					"Gender Bender",
-					"Harem",
-					"Historical",
-					"Horror",
-					"Josei",
-					"Martial Arts",
-					"Mature",
-					"Mecha",
-					"Mystery",
-					"Psychological",
-					"Romance",
-					"School Life",
-					"Sci-fi",
-					"Seinen",
-					"Shoujo",
-					"Shounen",
-					"Shounen Ai",
-					"Slice of Life",
-					"Smut",
-					"Sports",
-					"Supernatural",
-					"Tragedy",
-					"Wuxia",
-					"Xianxia",
-					"Xuanhuan",
-					"Yaoi",
-					"Yuri"
-				],
 				"values": [
 					"Action",
 					"Adult",
@@ -98,7 +61,6 @@
 				"type": "multi-select",
 				"key": "hiddenLanguages",
 				"title": "Hidden Original Languages",
-				"titles": ["Chinese", "Korean", "English", "Japanese"],
 				"values": ["Chinese", "Korean", "English", "Japanese"],
 				"default": [],
 				"refreshes": ["content"]
